Renames dashboard service
